MOTOR-40 Collect metaprogramming code into one module.
