+++ motor/motor_common.py
@@ -16,4 +16,30 @@ from __future__ import unicode_literals, absolute_import
 
 """Common code to support all async frameworks."""
 
+import pymongo.errors
+
+
 callback_type_error = TypeError("callback must be a callable")
+
+
+def check_deprecated_kwargs(kwargs):
+    if 'safe' in kwargs:
+        raise pymongo.errors.ConfigurationError(
+            "Motor does not support 'safe', use 'w'")
+
+    if 'slave_okay' in kwargs or 'slaveok' in kwargs:
+        raise pymongo.errors.ConfigurationError(
+            "Motor does not support 'slave_okay', use read_preference")
+
+    if 'auto_start_request' in kwargs:
+        raise pymongo.errors.ConfigurationError(
+            "Motor does not support requests")
+
+
+def mangle_delegate_name(motor_class, name):
+    if name.startswith('__') and not name.endswith("__"):
+        # Mangle, e.g. Cursor.__die -> Cursor._Cursor__die
+        classname = motor_class.__delegate_class__.__name__
+        return '_%s%s' % (classname, name)
+    else:
+        return name
